A whole bunch of new stuff:
run.sh: A script for passing a whole directory full of files as parameters to a program. sed magic in use.
bench.sh: A script to automate the benchmarking of code.
bank.hpp: A modified version of my data store.
datastore.hpp: A copy of Trav's data store (singly linked list).
disktest.cpp: A program that'll read and spew out parsed DNS data. Was going to be used for something else, but now is modified to work well with bench.sh
index.hpp: Primary file for index tables.
indextest.cpp: test file that benchmarks the index tables and data stores.
odb.cpp: Original test program for the original version of the data structure found in bank.hpp
linkedlist.hpp: Contains a linked list for use in an index table.