Moved all testing from minimal.cpp to test.cpp. Added an include to common.hpp. Other stuff...
linkedlisti.hpp : Made it such that teh verify() member exists only when debugging. Since it is the only member that requires stdio.h, that is included in that member. Not the standard way to do it, but it makes it easy to clean up.
odb.hpp : The switch for creating an index now uses the FAIL macro and I cleaned up the includes to adjust for the fact that it no longer needs to include any IO.
test.cpp : Rearranged as the datastore.hpp no long does what it used to. I removed that test case and changed the command line argument structure. Also added in repeat test functionality for more proper benchmarking.
	Usage: test <element size> <number of elements> <number of tests> <test type>
		where: test_type=0 => BANK, test_type=1 => LL

	Note: Forgot to change the usage() member.

Because test.cpp now includes proper benchmarking, bench.sh and test.sh ahve been deprecated and removed from future revisions.