BROKEN!: Somewhat ill-fated attempt (And verifiable waste of time, in hind sight) at a deferred-processing RBT.
It would defer tree manipulation until after a node was added. This would accelerate the process of handling duplicates and so be more efficient with datasets with a large amount of redundancy.

I was attempting this because I believe this is where the rbtree implementation gets its performance boost from: its algorithm is inherently two-pass. I attempted to adapt a one-pass algorithm into two-pass.

DOESN'T ACTUALLY WORK PROPERLY RIGHT NOW UNLESS COMPILED WITH -DLINUX_RBT (Which bypasses my modifications).