Implemented a prune function pointer passed as a parameter to all DataStore constructors.
Scaffolding for ODB added as well, and I would call it ready for use.
This parameter is mandatory and cannot be NULL. (TODO: Add some FAILs if it is). Becuse it can't be NULL, I added some (currently public) static functions in DataStore (prune_true and prune_false) that always return the respectiv value. (TODO: Make these private, force user to make their own. The ones in DataStore are for 'internal' use only).