Implemented proper remove_cleanup functions for all DS/IDS datastores. Works according to verify.sh.
DataStore::remove_sweep now returns an array of two vectors. The first element in the array are the addresses to remove from the index tables, and the second element contains the addresses to remove from the datastore when DataStore::remove_cleanup is called.

This is particularly useful for LinkedList datastores, as it allows for unlinking of the nodes without a O(n) operation. For BankDS, this incurs only 8 bytes of extra overhead, since the two vectors are identical. For the IDS datastores, this also provides a benefit.

All around, I think it is a good solution to the problem.