Added a scripts directory, and a processing script.
Moved the scripts out of "archive" and into a dedicated "scripts" folder, for cleanliness.

Also added a "process.sh" script that takes the output of demo-dns and converts it into eight files that can be read in by Mathematica as lists. One file for each of Ratio, Entropy, Total Packets, Unique Packets for the invalid and valid streams.