RedBlackTreeI now allows for embedded use. demo-protosim added.
Demo-protosim is a simulation of a TCP stack that will track connections. It makes use of the additional functionality added to the red black
tree that allows for use of an embedded tree.

An embedded tree is useful for small data items, or data items that will only ever have one index built on them.

The tcp-sim.*pp files were tracked, but are thus far empty (or at least useless). They'll be populated as demo-protosim matures.

The API for the buffered reader has changed slightly to allow for future use of a buffered writer. This can be taken advantage of in several
locations, such as the "AppendOnly" style of archiving removed data to disk. Explicit buffering in that situation would be beneficial.

Astyle-ing was done on the code base.