Added support to demo-driver for interval statistics.
At the moment it outputs the timestamp that marks the START of the interval, the entropy of the valid DNS queries in the interval, the total number of DNS queries over the interval, and the number of unique queries over the interval.

Next thing to do is handle invalid DNS over UDP53, and to change the pcap filter in the pcap handler. Possibly add an option to specify the pcap filter on the command line.