First run at a complete scheduler.
scripts/verify.sh passes on my Solaris box, but I haven't investigated the intermediate files much to see if they actually pass.

This will require a few revisions to work out the kinks, but the brunt of the work is done now.

The scheduler uses a hash_map for queue lookups when inserting work. These take care of most of the hard stuff, and are included with GCC 4.3+ and Sun Studio compilers. THis should work with any moderately recent Sun Studio compiler, as it uses the SGI STL hash_map which is built into the Sun Studio compiler library "stlport4". GCC however needs to make use of the unordered_map included with C++ TR1. This means that GCC and Sun compiled versions use fundamentally different code for the hash table, and this might result in differing performance characteristics for that portion of the processing.