Added some additional output to schedtest that prints out how many 'spins' were performed.
This allows a comparison of speed that is tempered by how long each one spent spinning to simulate a larger workload. test4() spins on every comparison, which simulates a complex comparison function, which is not the same as an insertion into a large tree. Both, however, are the same when looking at the workload duration.

The default values hard-coded into schedtest represent approximately equal workload sizes in each test.