Shuffled the whole directory structure around.
Moved the source code and includes into a "src" directory. Tests are moved into a "tests" directory, as well as the verify script. I've added a "make check" target that compiles and runs the tests. The tests aren't build as part of the "all" target. Application binaries are stored in a "bin" directory, and the libodb.so file is stored in the "src" directoyr with the other source files.