Cleaned out #ifdef/#if branching in headers, by replacing the branched members with opaque pointers.
All of the branching is now done in the .cpp files.

This necessitated a massive change to how lock.hpp works. Now you include it into your source file, as opposed to the header.
I haven't gotten around to porting all of the pthread/Google function macros, so things won't compile on Unix until I do.